manual review tidy up
